requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.1.4
Flask==0.12.2
idna==2.6
inflection==0.3.1
itsdangerous==0.24
Jinja2==2.10
lxml==4.1.1
MarkupSafe==1.0
more-itertools==4.1.0
ndg-httpsclient==0.4.4
newsapi-python==0.0.2
nltk==3.2.5
numpy==1.14.1
pandas==0.22.0
pandas-datareader==0.6.0
pyasn1==0.4.2
pycparser==2.18
pyOpenSSL==17.5.0
python-dateutil==2.6.1
pytz==2018.3
Quandl==3.3.0
requests==2.13.0
requests-file==1.4.3
requests-ftp==0.3.1
six==1.11.0
textblob==0.15.1
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
